Упростил структуру
