+++ app/src/main/java/root/iv/cocomoapp/ui/fragment/cocomo/visual/finance/VisualFinanceFragment.java
@@ -1,11 +1,7 @@
 package root.iv.cocomoapp.ui.fragment.cocomo.visual.finance;
 
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
 
-import com.github.mikephil.charting.charts.PieChart;
 import com.github.mikephil.charting.components.Description;
 import com.github.mikephil.charting.data.PieData;
 import com.github.mikephil.charting.data.PieDataSet;
@@ -15,41 +11,10 @@ import com.github.mikephil.charting.utils.ColorTemplate;
 import java.util.LinkedList;
 import java.util.List;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import root.iv.cocomoapp.R;
 import root.iv.cocomoapp.cocomo.ProjectParam;
-import root.iv.cocomoapp.ui.adapter.ProjectParamAdapter;
 import root.iv.cocomoapp.ui.fragment.cocomo.visual.BaseVisualFragment;
 
 public class VisualFinanceFragment extends BaseVisualFragment {
-    @BindView(R.id.listFinance)
-    protected RecyclerView listFinance;
-    @BindView(R.id.diagram)
-    protected PieChart pieFinanceDiagram;
-    private ProjectParamAdapter financeAdapter;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_visual_finance, container, false);
-        ButterKnife.bind(this, view);
-
-        Bundle args = getArguments();
-        mans = args.getDouble(ARG_MANS);
-        time = args.getDouble(ARG_TIME);
-
-        financeAdapter = new ProjectParamAdapter(getLayoutInflater());
-        listFinance.setAdapter(financeAdapter);
-        listFinance.setLayoutManager(new LinearLayoutManager(this.getContext(), LinearLayoutManager.HORIZONTAL, false));
-
-        fillFinance();
-        return view;
-    }
 
     public static VisualFinanceFragment getInstance(double man, double time) {
         VisualFinanceFragment fragment = new VisualFinanceFragment();
@@ -61,7 +26,8 @@ public class VisualFinanceFragment extends BaseVisualFragment {
         return fragment;
     }
 
-    public void fillFinance() {
+    @Override
+    public void fillContent() {
         List<ProjectParam> fin = new LinkedList<>();
 
         fin.add(ProjectParam.getInstance("Анализ требований (4%)", 0.04 * mans));
@@ -74,7 +40,7 @@ public class VisualFinanceFragment extends BaseVisualFragment {
         fin.add(ProjectParam.getInstance("Создание руководств (6%)", 0.06*mans));
 
         // Загружаем данные в адаптер
-        financeAdapter.append(fin);
+        adapter.append(fin);
 
 
         List<PieEntry> entrys = new LinkedList<>();
@@ -87,12 +53,16 @@ public class VisualFinanceFragment extends BaseVisualFragment {
 
         PieData data = new PieData(dataSet);
 
-        pieFinanceDiagram.setData(data);
+        diagram.setData(data);
         dataSet.setColors(ColorTemplate.MATERIAL_COLORS);
-        pieFinanceDiagram.animateXY(5000, 5000);
+        diagram.animateXY(5000, 5000);
         Description desc = new Description();
         desc.setText("Распределение ресурсов на разработку проекта");
-        pieFinanceDiagram.setDescription(desc);
+        diagram.setDescription(desc);
     }
 
+    @Override
+    public String getName() {
+        return "FINANCE";
+    }
 }
