+++ app/src/main/java/root/iv/cocomoapp/ui/fragment/cocomo/visual/lifecicle/VisualWorkLifecycleFragment.java
@@ -1,11 +1,7 @@
 package root.iv.cocomoapp.ui.fragment.cocomo.visual.lifecicle;
 
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
 
-import com.github.mikephil.charting.charts.PieChart;
 import com.github.mikephil.charting.components.Description;
 import com.github.mikephil.charting.data.PieData;
 import com.github.mikephil.charting.data.PieDataSet;
@@ -15,40 +11,10 @@ import com.github.mikephil.charting.utils.ColorTemplate;
 import java.util.LinkedList;
 import java.util.List;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import root.iv.cocomoapp.R;
 import root.iv.cocomoapp.cocomo.ProjectParam;
-import root.iv.cocomoapp.ui.adapter.ProjectParamAdapter;
 import root.iv.cocomoapp.ui.fragment.cocomo.visual.BaseVisualFragment;
 
 public class VisualWorkLifecycleFragment extends BaseVisualFragment {
-    @BindView(R.id.listParams)
-    protected RecyclerView listParams;
-    @BindView(R.id.diagram)
-    protected PieChart diagram;
-    private ProjectParamAdapter adapter;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_visual_work_lifecycle, container, false);
-        ButterKnife.bind(this, view);
-
-        Bundle args = getArguments();
-        mans = args.getDouble(ARG_MANS);
-        time = args.getDouble(ARG_TIME);
-
-        adapter = new ProjectParamAdapter(getLayoutInflater());
-        listParams.setAdapter(adapter);
-        listParams.setLayoutManager(new LinearLayoutManager(this.getContext(), RecyclerView.HORIZONTAL, false));
-        fillContent();
-        return view;
-    }
 
     public static VisualWorkLifecycleFragment getInstance(double m, double t) {
         VisualWorkLifecycleFragment fragment = new VisualWorkLifecycleFragment();
@@ -60,6 +26,7 @@ public class VisualWorkLifecycleFragment extends BaseVisualFragment {
         return fragment;
     }
 
+    @Override
     public void fillContent() {
         List<ProjectParam> params = new LinkedList<>();
 
@@ -86,5 +53,10 @@ public class VisualWorkLifecycleFragment extends BaseVisualFragment {
         Description desc = new Description();
         desc.setText("Распределение ресурсов на разработку проекта");
         diagram.setDescription(desc);
+    }
+
+    @Override
+    public String getName() {
+        return "WORK LIFECYCLE";
     }
 }
